feat:move mobile apis to application and ui module
