+++ python/agentbay/__init__.py
@@ -1,6 +1,11 @@
 from .agentbay import AgentBay
 from .exceptions import AgentBayError, APIError, AuthenticationError
 from .session import Session
-from .mobile import MobileSystem, KeyCode
+from .ui import UI
+from .oss import Oss
+from .filesystem import FileSystem
+from .window import Window
+from .command import Command
 
-__all__ = ["AgentBay", "Session", "AgentBayError", "AuthenticationError", "APIError", "MobileSystem", "KeyCode"]
+
+__all__ = ["AgentBay", "Session", "AgentBayError", "AuthenticationError", "APIError", "UI", "Oss", "FileSystem", "Window", "Command"]
