+++ python/agentbay/ui/__init__.py
@@ -1,0 +1,4 @@
+"""UI module for AgentBay."""
+from .ui import UI, KeyCode
+
+__all__ = ["UI", "KeyCode"]
